+++ src/components/sidebar/SidebarHeader.tsx
@@ -47,14 +47,14 @@ export default function SidebarHeader() {
             <IconSettings size={18} className="mr-1" />
             <span>Settings</span>
           </DropdownItem>
-          <DropdownItem
+          {/* <DropdownItem
             className="border-t dark:border-gray-700"
             as='link'
             href='https://pad.mdsilo.com'
           >
             <IconArtboard size={18} className="mr-1" />
             <span>Collaboration</span>
-          </DropdownItem>
+          </DropdownItem> */}
           <DropdownItem
             className="border-t dark:border-gray-700"
             as='link'