+++ src/components/sidebar/SidebarHistory.tsx
@@ -1,4 +1,4 @@
-import React, { memo, useCallback } from 'react';
+import React, { memo, useCallback, useEffect } from 'react';
 import { IconPin, IconTrash } from '@tabler/icons';
 import { useStore } from 'lib/store';
 import Tooltip from 'components/misc/Tooltip';
@@ -23,6 +23,17 @@ function SidebarHistory(props: Props) {
       await listInitDir(dir);
     }, []
   );
+
+  const isOpenPreOn = useStore((state) => state.isOpenPreOn);
+  const showHistory = useStore((state) => state.showHistory);
+  let isOpened = false;
+  useEffect(() => { 
+    console.log("open on", isOpenPreOn, "show history", showHistory, history)
+    if (!isOpenPreOn || showHistory || isOpened) return;
+    const recentDirPath = history.length > 0 ? history[history.length - 1] : '';
+    if (!recentDirPath) return;
+    listInitDir(recentDirPath).then(() => {isOpened = true; console.log("open previous folder")}); 
+  }, [history]);
 
   return (
     <div className={`flex flex-col flex-1 overflow-x-hidden ${className}`}>